Controller V0.2

+++ src/main/java/com/example/demo/Service/EtudiantService.java
@@ -24,7 +24,7 @@ public class EtudiantService implements IEtudiantService{
     }
 
     @Override
-    public Etudiant retrieveEtudiant(Long idEtudiant) {
+    public Etudiant retrieveEtudiant(Integer idEtudiant) {
         return etudiantRepository.findById(idEtudiant).get();
     }
